Updated endpoint names and configured Exception handling to be global across endpoints
