+++ src/main/kotlin/nomic/api/GlobalExceptionHandler.kt
@@ -1,29 +1,16 @@
 package nomic.api
 
 import nomic.data.EntityNotFoundException
-import nomic.domain.entities.RulesAmendmentsModel
-import nomic.domain.rulesamendments.RuleAmendmentDomain
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseStatus
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler
 
-@RestController
-@RequestMapping("api/rules_amendments")
-class RulesAmendments(val ruleAmendmentDomain: RuleAmendmentDomain) {
-    // Path to endpoint is api/rules_amendments/ExistingGameId
-    @GetMapping("{gameid}", produces = ["application/json;charset=UTF-8"])
-    fun getRulesAmendments(@PathVariable(value = "gameid") gameId: String): ResponseEntity<Any> {
-        val rulesAmendments: List<RulesAmendmentsModel> = ruleAmendmentDomain.getRulesAmendments(gameId)
 
-        // Return the response object
-        return ResponseEntity(rulesAmendments, HttpStatus.OK)
-    }
-
+@ControllerAdvice
+class GlobalExceptionHandler : ResponseEntityExceptionHandler() {
     @ExceptionHandler(IllegalArgumentException::class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     fun handleIllegalArgumentException(
