+++ src/main/kotlin/nomic/api/RulesAmendmentsEndpoint.kt
@@ -1,0 +1,23 @@
+package nomic.api
+
+import nomic.domain.entities.RulesAmendmentsModel
+import nomic.domain.rulesamendments.RuleAmendmentDomain
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("api/rules_amendments")
+class RulesAmendmentsEndpoint(val ruleAmendmentDomain: RuleAmendmentDomain) {
+    // Path to endpoint is api/rules_amendments/ExistingGameId
+    @GetMapping("{gameid}", produces = ["application/json;charset=UTF-8"])
+    fun getRulesAmendments(@PathVariable(value = "gameid") gameId: String): ResponseEntity<Any> {
+        val rulesAmendments: List<RulesAmendmentsModel> = ruleAmendmentDomain.getRulesAmendments(gameId)
+
+        // Return the response object
+        return ResponseEntity(rulesAmendments, HttpStatus.OK)
+    }
+}
